Let people disable gradebook client auth if they really want to
This adds a --no-auth command line argument that can be used to disable
the gradebook auth prompts. If that command line is passed, then anyone
will be able to connect to the gradebook server, without needing the
grader to allow it thru the grader CLI.

This is a lot of power; use it wisely (i.e. only when --host=localhost).