command line arguments
I was going to complicate the shit out of this by adding a command line
argument to take in a file containing more command line arguments... but
then I decided against it. It's not our job to replace the job of a
shell (and the glorious world of shell scripting, which already solves
this problem much better than we could've).